Init Android project. Empty.
