+++ app/src/main/java/com/poemgen/mockspire/poemgenerator/ai/PoemGenerator.kt
@@ -1,0 +1,6 @@
+package com.poemgen.mockspire.poemgenerator.ai
+
+interface PoemGenerator {
+    fun submitPrompt(prompt: String)
+    fun getPoem(): String
+}
