+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Thu Apr 29 11:31:28 NZST 2021
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-6.5-bin.zip
