--- poemgenerator/ai/PoemGenerator.kt
@@ -1,6 +1,0 @@
-package poemgenerator.ai
-
-interface PoemGenerator {
-    fun submitPrompt(prompt: String)
-    fun getPoem(): String
-}
