+++ stuff.txt
@@ -1,0 +1,3 @@
+lorem
+ipsum
+dolores
